+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -79,6 +79,8 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../util/logger.js",
   ],
   memory: [
+    "../../../../../agent/image-optimize.js",
+    "../../../../../config/loader.js",
     "../../../../../config/memory-v3-gate.js",
     "../../../../../config/types.js",
     "../../../../../messaging/providers/slack/message-metadata.js",
@@ -95,7 +97,6 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../../persistence/raw-query.js",
     "../../../../../persistence/schema/index.js",
     "../../../../../security/untrusted-content.js",
-    "../../../../agent/image-optimize.js",
     "../../../../api/events/memory-recalled.js",
     "../../../../api/responses/memory-v3-selection-log.js",
     "../../../../channels/types.js",
@@ -131,7 +132,6 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../persistence/embeddings/sparse-tokenize.js",
     "../../../../persistence/job-utils.js",
     "../../../../persistence/jobs-store.js",
-    "../../../../persistence/raw-query.js",
     "../../../../persistence/schema/conversations.js",
     "../../../../persistence/schema/index.js",
     "../../../../prompts/persona-resolver.js",
